modularise addLockedTokenBalances
