+++ client/src/containers/App/App.js
@@ -65,13 +65,10 @@ function App() {
     }
   }, [contractsLoaded, userAccount])
 
-  //TODO: change name from rewinder
   useEffect(() => {
     if (userFields.length && userTokens.length && !rewoundFlag) {
-
         apis.rewinder(userFields, trackedTokens, trackedFields)
           .then(rewound => {
-            console.log(' ---> rewound.userTokenBalances', rewound.userTokenBalances);
             setRewoundTokenBalances (prev => [...prev, ...rewound.userTokenBalances]);
             setRewoundFeederFieldBalances (prev => [...prev, ...rewound.userFeederFieldBalances]);
             setRewoundFlag(true);
@@ -81,35 +78,15 @@ function App() {
 
   //TODO: need to double-check this, figure out why flag is necessary and add feederFields
   //TODO: double-check trackedFields with fully populated crop/seedTokens
+  //TODO: why inconsistent populating meta?
   useEffect(() => {
 
     if (rewoundFlag) {
-      const updatedUserTokens = [...userTokens];
-
-      console.log(' ---> rewoundTokenBalances', rewoundTokenBalances);
-      
-      rewoundTokenBalances.forEach(rewoundToken => {
-        console.log(' ---> rewoundToken', rewoundToken);
-        const existingUserToken = updatedUserTokens.find(userToken => userToken.tokenId === rewoundToken.token.tokenId);
-        //FIXME: really bad to be mutating cache directly
-        //TODO: modularise
-        if (existingUserToken && existingUserToken.lockedBalance) {
-          existingUserToken.lockedBalance.push({balance: rewoundToken.userTokenBalance, field: rewoundToken.field});
-        }
-        else if (existingUserToken) existingUserToken.lockedBalance = [{balance: rewoundToken.userTokenBalance, field: rewoundToken.field}];
-        else {
-          const newUserToken = JSON.parse(JSON.stringify(rewoundToken.token));
-          newUserToken.lockedBalance = [{balance: rewoundToken.userTokenBalance, field: rewoundToken.field}]
-          updatedUserTokens.push(newUserToken);
-        }
-      })
-
-      setUserTokens(() => [...updatedUserTokens]);
+      const updatedUserTokens = helpers.addLockedTokenBalances(rewoundTokenBalances, userTokens);
+      setUserTokens(updatedUserTokens);
     }
 
   }, [rewoundTokenBalances, rewoundFlag])
-
-
 
   return (
     <div>
